Android: Fix for 64bit arm and x86_64 targets
tested on kitkat arm and lollipop x86_64/arm64

+++ android/librpc/bindresvport.c
@@ -38,7 +38,7 @@
 #include <string.h>
 #ifdef ANDROID
 #include <rpc/types.h>
-extern int __set_errno(int n);
+extern long __set_errno(int n);
 #else
 #include <sys/types.h>
 #endif
